+++ lsh/src/utils.cc
@@ -25,11 +25,16 @@ void utils::InputInfo::Print(void) {
   std::cout << "Input file: " << input_file << std::endl;
   std::cout << "Query file: " << query_file << std::endl;
   std::cout << "Output file: " << output_file << std::endl;
-  std::cout << "Number of LSH hash functions: " << K << std::endl;
-  std::cout << "Number of LSH hash tables: " << L << std::endl;
-  std::cout << "Number of dataset points: " << N << std::endl;
-  std::cout << "Number of query points: " << Q << std::endl;
-  std::cout << "Space dimension: " << D << std::endl;
+  std::cout << "Number of LSH hash functions: "
+            << static_cast<unsigned int>(K) << std::endl;
+  std::cout << "Number of LSH hash tables: "
+            << static_cast<unsigned int>(L) << std::endl;
+  std::cout << "Number of dataset points: "
+            << static_cast<unsigned int>(N) << std::endl;
+  std::cout << "Number of query points: "
+            << static_cast<unsigned int>(Q) << std::endl;
+  std::cout << "Space dimension: " 
+            << static_cast<unsigned int>(D) << std::endl;
   std::cout << "Radius: " << R << std::endl;
 }
